# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckInCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "medication": {
            "$ref": "#/components/schemas/CheckInMedication"
          },
          "pain": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CheckInPainRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "medication"
        ],
        "title": "CheckInCreateRequest",
        "type": "object"
      },
...
      "CheckInReadResponse": {
        "additionalProperties": false,
        "properties": {
          "answered": {
            "title": "Answered",
            "type": "boolean"
          },
          "answers": {
            "additionalProperties": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CheckInAnswerContent"
                },
                {
                  "type": "null"
                }
              ]
            },
            "propertyNames": {
              "$ref": "#/components/schemas/CheckInMedication"
            },
            "title": "Answers",
            "type": "object"
          },
          "demo_only": {
            "const": true,
            "default": true,
            "title": "Demo Only",
            "type": "boolean"
          },
          "drug_name": {
            "title": "Drug Name",
            "type": "null"
          },
          "next_checkin": {
            "title": "Next Checkin",
            "type": "null"
          },
          "next_visit": {
            "title": "Next Visit",
            "type": "null"
          },
          "pain_types": {
            "items": {
              "$ref": "#/components/schemas/CheckInPainTypeResponse"
            },
            "title": "Pain Types",
            "type": "array"
          },
          "round_label": {
            "const": "복약 7일째 · 첫 확인",
            "default": "복약 7일째 · 첫 확인",
            "title": "Round Label",
            "type": "string"
          }
        },
        "required": [
          "answers",
          "pain_types",
          "answered"
        ],
        "title": "CheckInReadResponse",
        "type": "object"
      },
...
      "GuidePageViewRequest": {
        "additionalProperties": false,
        "description": "환자가 연 장 하나 — KEY-256.\n\n**칸이 하나뿐이다.** 언제 열었는지는 서버가 찍고, 누가 열었는지는 토큰이\n말한다. 화면이 시각을 보내면 브라우저 시계를 믿는 셈이 되고, 이 저장소는\n그 부류로 이미 크게 데었다(저장 시각이 아홉 시간 어긋나 링크 만료 비교가\n깨져 있었다).",
        "properties": {
          "section": {
            "$ref": "#/components/schemas/GuideSectionKey"
          }
        },
        "required": [
          "section"
        ],
        "title": "GuidePageViewRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VisitListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VisitResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/CursorPage"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "VisitListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/checkins/{token}": {
      "get": {
        "operationId": "read_patient_checkin_api_v1_checkins__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "patient_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInReadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Patient Checkin",
        "tags": [
          "patient-checkins"
        ]
      },
      "post": {
        "operationId": "save_patient_checkin_api_v1_checkins__token__post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "patient_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInSaveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Patient Checkin",
        "tags": [
          "patient-checkins"
        ]
      }
    },
...
    "/api/v1/guides/{token}": {
      "get": {
        "operationId": "read_patient_guide_api_v1_guides__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "patient_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientGuideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Patient Guide",
        "tags": [
          "patient-guides"
        ]
      }
    },
    "/api/v1/guides/{token}/views": {
      "post": {
        "description": "환자가 안내문의 **한 장**을 열었다 — KEY-256, 원문 S2-2·D1-6.\n\n링크로 들어오는 `GET /{token}` 은 안내문을 통째로 내려 주고, 환자 화면은\n탭을 **브라우저 안에서만** 넘긴다. 그래서 서버에는 「열었다」한 줄만 남고\n「어디까지 읽었나」가 없었다. 스탭은 다 읽은 환자와 첫 장만 열고 닫은\n환자를 구별할 수 없었는데, 다음 진료 때 물을 것이 서로 다르다.\n\n**본문에 원문이 없다.** 어느 장을 열었는지 하나뿐이다 — 이 저장소가\n이용 기록에 두는 선(`app/services/patient_usage.py`)을 그대로 지킨다.\n\n**`204` 로 답한다.** 환자 화면이 받을 것이 없다. 그리고 이 호출이 실패해도\n화면은 계속 읽혀야 한다 — 통계가 환자의 안내 열람을 막으면 안 된다.\n같은 이유로 화면 쪽은 이 호출을 기다리지 않는다.",
        "operationId": "record_guide_page_view_api_v1_guides__token__views_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidePageViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Guide Page View",
        "tags": [
          "patient-guides"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}": {
      "get": {
        "operationId": "get_patient_api_v1_patients__patient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Patient",
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "update_patient_api_v1_patients__patient_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Patient",
        "tags": [
          "patients"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/visits": {
      "get": {
        "operationId": "list_visits_api_v1_patients__patient_id__visits_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Visits",
        "tags": [
          "visits"
        ]
      },
      "post": {
        "operationId": "create_visit_api_v1_patients__patient_id__visits_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Visit",
        "tags": [
          "visits"
        ]
      }
    },
...
    "/api/v1/visits/{visit_id}": {
      "get": {
        "operationId": "get_visit_api_v1_visits__visit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Visit",
        "tags": [
          "visits"
        ]
      },
      "patch": {
        "operationId": "update_visit_api_v1_visits__visit_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Visit",
        "tags": [
          "visits"
        ]
      }
    },
...
    "/api/v1/visits/{visit_id}/checkin": {
      "get": {
        "operationId": "read_hospital_checkin_api_v1_visits__visit_id__checkin_get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HospitalCheckInResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Hospital Checkin",
        "tags": [
          "patient-links"
        ]
      }
    },
...
    "/api/v1/visits/{visit_id}/guide/link": {
      "delete": {
        "description": "병원에서 현재 환자 링크를 즉시 폐기한다 — KEY-223.",
        "operationId": "revoke_patient_guide_link_api_v1_visits__visit_id__guide_link_delete",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Patient Guide Link",
        "tags": [
          "patient-links"
        ]
      },
      "get": {
        "description": "링크 상태를 읽는다 — KEY-275. **주소는 안 나간다.**\n\n두 화면(문자 설정 S1-14 · 현황 D1-6)이 같은 블록을 그리는데, 지금까지는\n링크가 살아 있는지 볼 길이 아예 없었다. 현황 화면은 안내문 API 를 안\n부르므로 `GuideResponse` 에 얹지 않고 이 길을 따로 냈다.",
        "operationId": "read_patient_guide_link_state_api_v1_visits__visit_id__guide_link_get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientLinkStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Patient Guide Link State",
        "tags": [
          "patient-links"
        ]
      },
      "post": {
        "operationId": "issue_patient_guide_link_api_v1_visits__visit_id__guide_link_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientLinkIssueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Issue Patient Guide Link",
        "tags": [
          "patient-links"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/guide/link/re-issue": {
      "post": {
        "description": "병원에서 링크를 교체하고 새 원문을 이번 응답에만 돌려준다 — KEY-223.",
        "operationId": "re_issue_patient_guide_link_api_v1_visits__visit_id__guide_link_re_issue_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientLinkIssueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Re Issue Patient Guide Link",
        "tags": [
          "patient-links"
        ]
      }
    },
...
  }
}